# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ReferralTierId =
  (typeof ReferralTierId)[keyof typeof ReferralTierId];

export const ReferralTierId = {
  recruit: "recruit",
  operator: "operator",
  captain: "captain",
  commander: "commander",
  general: "general",
} as const;

/**
 * A rung on the recruiter ladder (social-only, no entitlements).
 */
export interface ReferralTier {
  id: ReferralTierId;
  /** Display label (e.g. "Captain") */
  label: string;
  /**
   * Minimum total claims to reach this tier
   * @minimum 0
   */
  claimsRequired: number;
}

export interface ReferralInfo {
  /**
   * User's stable invite code (8 chars, ambiguity-safe alphabet)
   * @minLength 6
   * @maxLength 16
   */
  code: string;
  /**
   * Count of users who have signed up using this code
   * @minimum 0
   */
  totalClaims: number;
  /** Anonymous public handle derived from the invite code (e.g. "Operator GQ55"). Stable per user, zero PII.
   */
  handle: string;
  tier: ReferralTier;
  nextTier?: ReferralTier;
  /**
   * Claims remaining to reach next tier; 0 at top tier
   * @minimum 0
   */
  claimsToNextTier: number;
}

export interface ReferralLeaderboardEntry {
  handle: string;
  tier: ReferralTier;
  /** @minimum 0 */
  claims: number;
  /** @minimum 1 */
  rank: number;
  isYou: boolean;
}

export interface ReferralLeaderboard {
  entries: ReferralLeaderboardEntry[];
  /**
   * 1-indexed rank; 0 means unranked (no claims yet)
   * @minimum 0
   */
  yourRank: number;
  /** @minimum 0 */
  yourClaims: number;
  /** @minimum 0 */
  totalParticipants: number;
}

export interface ReferralClaimInput {
  /**
   * Invite code shared by another user
   * @minLength 6
   * @maxLength 16
   */
  code: string;
}

export interface ReferralClaimResult {
  ok: boolean;
  referrerUserId: string;
}