# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: referrals
    description: Social-only referral codes & attribution (spec #7)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /referrals/me:
    get:
      operationId: getMyReferralInfo
      tags: [referrals]
      summary: Get my referral code and attribution stats
      description: >
        Returns the current user's invite code (auto-generated on first
        call) and the number of new users who have claimed it. Auth
        required.
      responses:
        "200":
          description: Referral info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /referrals/claim:
    post:
      operationId: claimReferral
      tags: [referrals]
      summary: Claim a referral code at signup
      description: >
        Records that the current user signed up using another user's
        invite code. One claim per user lifetime. Self-claims and
        unknown codes are rejected. No reward is granted (social-only
        leaderboard slice).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralClaimInput"
      responses:
        "200":
          description: Claim recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralClaimResult"
        "400":
          description: Bad request (malformed code or self-claim)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unknown referral code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already claimed a referral
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /referrals/leaderboard:
    get:
      operationId: getReferralLeaderboard
      tags: [referrals]
      summary: Top recruiters + the caller's rank
      description: >
        Returns the top recruiters by total claims (anonymous handles
        only — no PII) plus the caller's own rank if they have ranked.
        Auth required.
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralLeaderboard"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ReferralTier:
      type: object
      description: A rung on the recruiter ladder (social-only, no entitlements).
      properties:
        id:
          type: string
          enum: [recruit, operator, captain, commander, general]
        label:
          type: string
          description: Display label (e.g. "Captain")
        claimsRequired:
          type: integer
          minimum: 0
          description: Minimum total claims to reach this tier
      required:
        - id
        - label
        - claimsRequired
    ReferralInfo:
      type: object
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 16
          description: User's stable invite code (8 chars, ambiguity-safe alphabet)
        totalClaims:
          type: integer
          minimum: 0
          description: Count of users who have signed up using this code
        handle:
          type: string
          description: >
            Anonymous public handle derived from the invite code
            (e.g. "Operator GQ55"). Stable per user, zero PII.
        tier:
          $ref: "#/components/schemas/ReferralTier"
        nextTier:
          $ref: "#/components/schemas/ReferralTier"
        claimsToNextTier:
          type: integer
          minimum: 0
          description: Claims remaining to reach next tier; 0 at top tier
      required:
        - code
        - totalClaims
        - handle
        - tier
        - claimsToNextTier
    ReferralLeaderboardEntry:
      type: object
      properties:
        handle:
          type: string
        tier:
          $ref: "#/components/schemas/ReferralTier"
        claims:
          type: integer
          minimum: 0
        rank:
          type: integer
          minimum: 1
        isYou:
          type: boolean
      required:
        - handle
        - tier
        - claims
        - rank
        - isYou
    ReferralLeaderboard:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/ReferralLeaderboardEntry"
        yourRank:
          type: integer
          minimum: 0
          description: 1-indexed rank; 0 means unranked (no claims yet)
        yourClaims:
          type: integer
          minimum: 0
        totalParticipants:
          type: integer
          minimum: 0
      required:
        - entries
        - yourRank
        - yourClaims
        - totalParticipants
    ReferralClaimInput:
      type: object
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 16
          description: Invite code shared by another user
      required:
        - code
    ReferralClaimResult:
      type: object
      properties:
        ok:
          type: boolean
        referrerUserId:
          type: string
      required:
        - ok
        - referrerUserId