# lib/api-zod/src/generated/types/referralInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReferralTier } from "./referralTier";

export interface ReferralInfo {
  /**
   * User's stable invite code (8 chars, ambiguity-safe alphabet)
   * @minLength 6
   * @maxLength 16
   */
  code: string;
  /**
   * Count of users who have signed up using this code
   * @minimum 0
   */
  totalClaims: number;
  /** Anonymous public handle derived from the invite code (e.g. "Operator GQ55"). Stable per user, zero PII.
   */
  handle: string;
  tier: ReferralTier;
  nextTier?: ReferralTier;
  /**
   * Claims remaining to reach next tier; 0 at top tier
   * @minimum 0
   */
  claimsToNextTier: number;
}