# lib/api-zod/src/generated/types/referralLeaderboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReferralLeaderboardEntry } from "./referralLeaderboardEntry";

export interface ReferralLeaderboard {
  entries: ReferralLeaderboardEntry[];
  /**
   * 1-indexed rank; 0 means unranked (no claims yet)
   * @minimum 0
   */
  yourRank: number;
  /** @minimum 0 */
  yourClaims: number;
  /** @minimum 0 */
  totalParticipants: number;
}