# lib/api-zod/src/generated/types/referralLeaderboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReferralTier } from "./referralTier";

export interface ReferralLeaderboardEntry {
  handle: string;
  tier: ReferralTier;
  /** @minimum 0 */
  claims: number;
  /** @minimum 1 */
  rank: number;
  isYou: boolean;
}