# lib/api-zod/src/generated/types/referralTier.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReferralTierId } from "./referralTierId";

/**
 * A rung on the recruiter ladder (social-only, no entitlements).
 */
export interface ReferralTier {
  id: ReferralTierId;
  /** Display label (e.g. "Captain") */
  label: string;
  /**
   * Minimum total claims to reach this tier
   * @minimum 0
   */
  claimsRequired: number;
}